make sensor_reading_time_unix_micro unsigned long long for c ABI

+++ viam-cartographer/src/carto_facade/carto_facade.h
@@ -82,7 +82,7 @@ typedef struct viam_carto_sensor_reading {
     bstring sensor;
     // TODO: change to void* and a size
     bstring sensor_reading;
-    uint64_t sensor_reading_time_unix_micro;
+    unsigned long long sensor_reading_time_unix_micro;
 } viam_carto_sensor_reading;
 
 typedef enum viam_carto_LIDAR_CONFIG {
